requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
Django==4.2.1
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
